+++ affiliate/blocks/productCard/block.json
