+++ affiliate/blocks/productCard/src/OrganicIcon.jsx
