+++ affiliate/blocks/productCard/src/ProductCard.jsx
