+++ affiliate/blocks/productCard/src/productSearchModal.scss
