+++ affiliate/blocks/productCard/src/propTypes.js
