+++ affiliate/blocks/productCarousel/block.json
