+++ affiliate/blocks/productCarousel/src/ProductCarousel.jsx
