+++ affiliate/blocks/productCarousel/src/propTypes.js
