+++ affiliate/blocks/productCarousel/src/style.scss
@@ -5,17 +5,10 @@
  * Replace them with your own styles or remove the file completely.
  */
 
-.wp-block-organic-affiliate-product-card {
-  width: 100%;
-}
 .wp-block-organic-affiliate-product-carousel {
   width: 100%;
 }
 
-[data-organic-affiliate-integration="product-card"] {
-  min-height: 540px;
-  font-size: 0;
-}
 [data-organic-affiliate-integration="product-carousel"] {
   min-height: 580px;
   font-size: 0;
