+++ affiliate/shared/IntegrationModal.jsx
