+++ affiliate/shared/OrganicIcon.jsx
@@ -1,0 +1,16 @@
+const OrganicIcon = () => (
+    <svg
+        fill="none"
+        height={20}
+        viewBox="0 0 25 26"
+        width={20}
+    >
+        <path
+            clipRule="evenodd"
+            d="M0.260646 10.387C1.59838 3.14112 7.33154 -1.08564 14.3547 0.24277C21.0911 1.49872 25.0804 7.51279 23.7427 14.7828C22.4288 22.0045 16.5046 26.0622 9.57702 24.758C2.84056 23.4778 -1.07709 17.6329 0.260646 10.387ZM3.19889 9.80731C3.19889 5.87039 5.8027 2.27161 9.2187 1.03982C9.2187 1.03982 10.2937 0.605063 11.6075 0.74998C13.7097 0.967356 15.7163 2.175 17.054 3.69663C18.3918 5.21827 19.9206 8.45475 20.2311 11.5222C20.3745 12.8023 20.1356 15.7731 19.8012 17.1256C18.559 22.1011 16.0268 24.5889 12.9691 24.2024C10.9864 23.9609 9.88757 23.0431 9.12315 20.5312C8.52594 18.5507 7.97652 17.7053 6.37601 16.2561C4.05886 14.1307 3.19889 12.3675 3.19889 9.80731Z"
+            fillRule="evenodd"
+        />
+    </svg>
+);
+
+export default OrganicIcon;
