+++ affiliate/shared/integrationModal.scss
