--- affiliate/src/productCarousel/components/integrationModal.scss
@@ -1,9 +1,0 @@
-.integration-modal {
-  .components-modal__content {
-    padding: 0;
-
-    &::before {
-      margin: 0;
-    }
-  }
-}
